docker/etl/requirements.txt
requests==2.31.0
msal==1.24.1
pandas==2.1.3
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Office365-REST-Python-Client==2.5.3
